default_resolvers
