+++ modules/runtime/include/misc/types.h
@@ -210,12 +210,14 @@ BLOB_POD(skr_md5_t);
 
 namespace skr
 {
+using SInterfaceDeleter = void(*)(struct SInterface*);
 struct RUNTIME_API SInterface
 {
     virtual ~SInterface() = default;
     virtual uint32_t add_refcount() = 0;
     virtual uint32_t release() = 0;
     virtual skr_guid_t get_type() { return {}; }
+    virtual SInterfaceDeleter custom_deleter() const { return nullptr; }
 };
 template <class T>
 constexpr bool is_object_v = std::is_base_of_v<skr::SInterface, T>;
