+++ tools/tool_core/src/asset/cook_system.cpp
@@ -102,9 +102,7 @@ struct TOOL_CORE_API SkrToolCoreModule : public skr::IDynamicModule
                 desc.sort_method = SKR_ASYNC_SERVICE_SORT_METHOD_NEVER;
                 desc.sleep_mode = SKR_ASYNC_SERVICE_SLEEP_MODE_SLEEP;
                 ioService = skr_io_ram_service_t::create(&desc);        
-                ioService->add_file_resolver();
-                ioService->add_iobuffer_resolver();
-                ioService->add_chunking_resolver();
+                ioService->add_default_resolvers();
             }
         }
     }
